Déplace le menu des filtres dans un composant dédié
